requirements.txt
flask==2.3.3
Pillow==9.5.0
numpy==1.24.3
pytest==8.4.1